[Finishes Chore #161568013] Add response message and change folder structure
