+++ app/api/v2/views/users_endpoint.py
@@ -1,4 +1,4 @@
-from flask import Flask, jsonify, request
+from flask import Flask, jsonify, request, make_response
 from flask_restful import Resource
 from app import create_app
 from flask_jwt_extended import (
@@ -23,7 +23,7 @@ class UsersApi(Resource):
         """
         claims = get_jwt_claims()
         if claims['role'] != "admin":
-            return jsonify({'message':'You are not allowed to perform this action, contact the system admin!'})
+            return make_response(jsonify({'message':'You are not allowed to perform this action, contact the system admin!'}),401)
 
         data=request.get_json()
         if not data:
@@ -37,22 +37,21 @@ class UsersApi(Resource):
         user_details = [email,names,password,role]
         for field in user_details:
             if not field or field.isspace():
-                return jsonify({'message':'Some fields are empty!'})
+                return make_response(jsonify({'message':'Some fields are empty!'}),200)
 
         if not user_object.validate_email(email):
-            return jsonify({'message':'Please enter a valaid email'})
+            return make_response(jsonify({'message':'Please enter a valid email'}),200)
 
         if user_object.get_one_user(email):
-            return {'message':'User account alrleady exists'}
+            return make_response(jsonify({'message':'User account alrleady exists'}),200)
 
         if not (role == "admin" or role == "attendant"):
-            return jsonify({'message':'Roles can only be admin or attendant'})
+            return make_response(jsonify({'message':'Roles can only be admin or attendant'}),200)
 
         if not user_object.validate_password(password):
-            return jsonify({'message':'password should be between 6 and 12 characters, have atleast one lower_case,'+
-            'one Upper_case,one number and one special character'})
-        response = jsonify(user_object.add_user(email,names,password,role))
-        response.status_code = 201
+            return make_response(jsonify({'message':'password should be between 6 and 12 characters, have atleast one lower_case,'+
+            'one Upper_case,one number and one special character'}),200)
+        response = make_response(jsonify(user_object.add_user(email,names,password,role)),201)
         return response
     @jwt_required
     def get(self):
@@ -63,8 +62,8 @@ class UsersApi(Resource):
 
         users = user_object.get_all_users()
 
-        response = jsonify({"This are users in the system":users})
-        response.status_code = 200
+        response = make_response(jsonify({"This are users in the system":users}),200)
+
 
         return response
 
@@ -84,11 +83,10 @@ class SingleUserApi(Resource):
         password = data.get('password')
 
         if not user_object.verify_user(email,password):
-            return jsonify(dict(message = "wrong email or password"))
+            return make_response(jsonify({"message": "wrong email or password"}),401)
         logged_user = user_object.get_one_user(email)
         names = logged_user["names"]
         role = logged_user["role"]
         access_token = create_access_token(identity = logged_user)
-        response = jsonify(dict(token = access_token, message = "wellcome "+names +", "+"you are loged in as "+role))
-        response.status_code = 200
+        response = make_response(jsonify(dict(token = access_token, message = "wellcome "+names +", "+"you are loged in as "+role)),201)
         return response
